更新规则

+++ src/main/java/service/imp/standard/StandardCheckInf.java
@@ -16,10 +16,10 @@ public interface StandardCheckInf {
     //    查表或是代码中获取原始基础分数以及约束条件
 //    public Map getScoreAndExtremumFromTable(StandardDao standardDao, Map info);
     //    二次有效性检查，检查填写信息是否超出约束
-    Map isExtrrmumBand(Map map, int min, int max);
+    Map isExtrrmumBand(Map map, double min, double max);
 
     //    根据完整信息确定权重，并且根据文件决定分数是自动分配还是给出总分
-    Map getFinalScore(Map map, int tableScore);
+    Map getFinalScore(Map map, double tableScore);
 
     //    确认检测，考察分数分配是否符合规定
     Map confirmCheck(Map map, Map score);
